RF: Subclass Address from UnsignedInt
