+++ BitTornado/Tracker/track.py
@@ -22,11 +22,10 @@ from BitTornado.Client.Announce import HTTPAnnouncer, Response
 from BitTornado.Meta.bencode import bencode, Bencached, BencodedFile
 from BitTornado.Network.BTcrypto import CRYPTO_OK
 from BitTornado.Network.NatCheck import NatCheck, CHECK_PEER_ID_ENCRYPTED
-from BitTornado.Network.NetworkAddress import is_valid_ip, to_ipv4, AddrList, \
-    IPv4
+from BitTornado.Network.NetworkAddress import is_valid_ip, to_ipv4, AddrList
 from BitTornado.Network.RawServer import RawServer, autodetect_socket_style
 from ..Types import TypedDict, BytesIndexed, Infohash, PeerID, Port, \
-    UnsignedInt
+    UnsignedInt, IPv4
 from BitTornado.clock import clock
 
 from BitTornado import version
