+++ BitTornado/Types/__init__.py
@@ -1,6 +1,7 @@
 from .primitives import FixedLengthBytes, SixBytes, TwentyBytes, UnsignedInt, \
     UnsignedShort
 from .aliases import Infohash, PeerID, Port
+from .addresses import Address, IP, IPv4, IPv6
 from .bitfield import Bitfield, TrueBitfield
 from .collections import TypedList, SplitList, TypedDict, BytesIndexed, \
     QueryDict
