+++ BitTornado/Types/tests/__init__.py
@@ -1,3 +1,4 @@
 from .test_primitives import *
+from .test_addresses import *
 from .test_bitfield import *
 from .test_collections import *
