+++ BitTornado/tests/__init__.py
@@ -1,6 +1,6 @@
 from ..Types.tests import *
 from .test_bencode import CodecTests
-from .test_networkaddress import AddressTests, AddressRangeTests, \
+from .test_networkaddress import AddressFunctionTests, AddressRangeTests, \
     SubnetTests, TestAddrList
 from .test_parseargs import ParseArgsTest
 from .test_piecebuffer import PieceBufferTests
